refactor: delete unused files + renames
